+++ comments-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao comments bundle change log
 
+### DEV
+
+ * Notify subscribers of replies (see contao/core#8565).
+
 ### 4.4.0-beta1 (2017-05-05)
 
  * Optimize the element preview height (see contao/core-bundle#678).